// apps/envie-front/models/Replacement_piece/UpdateReplacementPiece.tsx
export interface UpdateReplacementPiece{
    state: string;
    // weight: number;
    subfamilyId: string;
    // excludingTaxesPrice: number;
    // indicativePriceOfNewPiece: number;
    comments: string;
    manufacturerReferenceOfPiece: string;
    designationInterne: string;
    brandId: string;
    model: string;
    technicalReferenceDevice: string;
    serialNumber: string;
  }